fix: align worker cancellation contracts and fixtures
Move dispatch admission contracts into the existing service contract and use
the canonical placement record to remove import cycles.

Update bootstrap preflight and late admission fixtures to observe the actual
owner lifetimes, and always settle held installation work before teardown.
