+++ src/gateway/server-worker-environment-startup.bootstrap.test.ts
@@ -158,9 +158,11 @@ describe("cloud bootstrap plugin generations", () => {
         await manager.prepare(startup.store.get("first")!);
         const preparationSignal = producers[0]!.prepare.mock.calls.at(-1)?.[0];
         expect(preparationSignal).toBeInstanceOf(AbortSignal);
-        expect(preparationSignal?.aborted).toBe(false);
+        expect(preparationSignal?.aborted).toBe(true);
+        expect(first.signal?.aborted).toBe(false);
+        expect(sameGeneration.signal?.aborted).toBe(false);
         manager.close(sameGeneration);
-        expect(preparationSignal?.aborted).toBe(false);
+        expect(first.signal?.aborted).toBe(false);
 
         registry = makeRegistry("runtime-b");
         const replacement = await begin("replacement");
@@ -188,7 +190,6 @@ describe("cloud bootstrap plugin generations", () => {
           stopped = true;
         });
         await producers[2]!.closing;
-        expect(preparationSignal?.aborted).toBe(true);
         expect(first.signal?.aborted).toBe(true);
         expect(replacement.signal?.aborted).toBe(true);
         expect(refreshed.signal?.aborted).toBe(true);
